capstones/therealtplum/apps/python-etl/requirements.txt
requests
psycopg[binary]
python-dotenv
psycopg2-binary